moving to simple implementation for command expansion
