--- .env.template
@@ -1,20 +1,0 @@
-AERIE_USERNAME=aerie_admin
-AERIE_PASSWORD=
-
-GATEWAY_USERNAME=gateway_user
-GATEWAY_PASSWORD=
-
-MERLIN_USERNAME=merlin_user
-MERLIN_PASSWORD=
-
-SCHEDULER_USERNAME=scheduler_user
-SCHEDULER_PASSWORD=
-
-SEQUENCING_USERNAME=sequencing_user
-SEQUENCING_PASSWORD=
-
-POSTGRES_USER=postgres_user
-POSTGRES_PASSWORD=
-
-HASURA_GRAPHQL_ADMIN_SECRET=
-HASURA_GRAPHQL_JWT_SECRET=
