+++ contrib/src/main/java/gov/nasa/jpl/aerie/contrib/streamline/modeling/Registrar.java
@@ -37,7 +37,7 @@ import static gov.nasa.jpl.aerie.merlin.framework.ModelActions.waitUntil;
  * </p>
  */
 public class Registrar {
-  private final gov.nasa.jpl.aerie.merlin.framework.Registrar baseRegistrar;
+  public final gov.nasa.jpl.aerie.merlin.framework.Registrar baseRegistrar;
   private boolean trace = false;
   private boolean profile = false;
   private final ErrorBehavior errorBehavior;
