+++ examples/command-expansion-modeling/src/main/java/gov/nasa/jpl/aerie/command_expansion_modeling/Configuration.java
@@ -1,0 +1,9 @@
+package gov.nasa.jpl.aerie.command_expansion_modeling;
+
+import gov.nasa.jpl.aerie.merlin.framework.annotations.Export;
+import gov.nasa.jpl.aerie.merlin.protocol.types.Duration;
+
+public final class Configuration {
+    @Export.Parameter
+    public Duration defaultCommandDuration = Duration.SECOND;
+}
