+++ examples/command-expansion-modeling/src/main/java/gov/nasa/jpl/aerie/command_expansion_modeling/activities/Activity_C.java
@@ -1,0 +1,4 @@
+package gov.nasa.jpl.aerie.command_expansion_modeling.activities;
+
+public class Activity_C {
+}
