+++ examples/command-expansion-modeling/src/main/java/gov/nasa/jpl/aerie/command_expansion_modeling/sequencing/Command.java
@@ -1,0 +1,9 @@
+package gov.nasa.jpl.aerie.command_expansion_modeling.sequencing;
+
+import gov.nasa.jpl.aerie.merlin.framework.annotations.AutoValueMapper;
+
+import java.util.List;
+import java.util.Map;
+
+@AutoValueMapper.Record
+public record Command(String stem, List<String> arguments) {}
