+++ examples/command-expansion-modeling/src/main/java/gov/nasa/jpl/aerie/command_expansion_modeling/sequencing/ExecutableCommand.java
@@ -1,0 +1,9 @@
+package gov.nasa.jpl.aerie.command_expansion_modeling.sequencing;
+
+import gov.nasa.jpl.aerie.command_expansion_modeling.activities.commands.CommandActivity;
+
+// TODO - I don't like the name "base" here.
+//  The name should reflect that we're separating the written representation of the command
+//  from its implementation
+public record ExecutableCommand(Command base, CommandActivity activity, CommandBehavior behavior) {
+}
