+++ examples/command-expansion-modeling/src/main/java/gov/nasa/jpl/aerie/command_expansion_modeling/sequencing/SubSequence.java
@@ -1,0 +1,10 @@
+package gov.nasa.jpl.aerie.command_expansion_modeling.sequencing;
+
+import gov.nasa.jpl.aerie.merlin.framework.annotations.AutoValueMapper;
+
+import java.time.Instant;
+import java.util.Map;
+
+@AutoValueMapper.Record
+public record SubSequence(Map<Instant,Command> commands) {
+}
