+++ examples/command-expansion-modeling/src/main/resources/META-INF/services/gov.nasa.jpl.aerie.merlin.protocol.model.SchedulerPlugin
@@ -1,0 +1,1 @@
+gov.nasa.jpl.aerie.command_expansion_modeling.generated.GeneratedSchedulerPlugin
