--- examples/command-model/src/main/java/gov/nasa/jpl/aerie/command_model/Configuration.java
@@ -1,9 +1,0 @@
-package gov.nasa.jpl.aerie.command_model;
-
-import gov.nasa.jpl.aerie.merlin.framework.annotations.Export;
-import gov.nasa.jpl.aerie.merlin.protocol.types.Duration;
-
-public final class Configuration {
-    @Export.Parameter
-    public Duration defaultCommandDuration = Duration.SECOND;
-}
