--- examples/command-model/src/main/java/gov/nasa/jpl/aerie/command_model/sequencing/Sequence.java
@@ -1,9 +1,0 @@
-package gov.nasa.jpl.aerie.command_model.sequencing;
-
-import gov.nasa.jpl.aerie.merlin.framework.annotations.AutoValueMapper;
-
-import java.util.List;
-
-@AutoValueMapper.Record
-public record Sequence(String id, List<Command> commands) {
-}
