--- examples/command-model/src/main/java/gov/nasa/jpl/aerie/command_model/sequencing/command_dictionary/ControlFlowDictionary.java
@@ -1,13 +1,0 @@
-package gov.nasa.jpl.aerie.command_model.sequencing.command_dictionary;
-
-import gov.nasa.jpl.aerie.command_model.sequencing.CommandBehavior;
-import gov.nasa.jpl.aerie.command_model.sequencing.Sequence;
-
-import java.util.List;
-
-/**
- * Specifies only the control flow, and no other part of the command behavior.
- */
-public interface ControlFlowDictionary {
-    List<CommandBehavior> interpret(Sequence sequence);
-}
