--- examples/command-model/src/main/resources/META-INF/services/gov.nasa.jpl.aerie.merlin.protocol.model.MerlinPlugin
@@ -1,1 +1,0 @@
-gov.nasa.jpl.aerie.command_model.generated.GeneratedMerlinPlugin
